Thirteen real driver sources survive in the laboratory, and fourteen drivers come back to life
24 eros_* drivers were inert for want of a SOURCE.json that existed nowhere
in the tree. Thirteen genuine ones survived, tracked, in the frozen
laboratory at src/soma/observations/, plus two result directories. All
recovered whole to soma/life/driver-sources/, each stamped with a provenance
member naming repo, commit and path.

14 OF 21 INPUT-BLOCKED DRIVERS NOW RUN. Thirteen from genuine recovery; one
from a declared port.

And eros_resonant_generation reproduces CLAUDE.md section 5's headline figure
live: 11,879 suffix states, 427 branches, 336 previously-unseen prefixes,
with the OLMo tokenizer's sha256 MATCHING the frozen source's declared
digest. The laboratory's flagship generation result is no longer a citation.

Also live: the pretrained ecology admitting 15 source factors beyond 6
inherited from OLMo BF16 as exact dyadics; the translation atlas where one
site RIDEs at one receiver and is OPEN at another; a 1.09 GB cohered-corpus
observation; the parent invoked only on the five OPEN paths of an
8,192-factor relation.

## What was refused rather than authored

The four measurement-blocked drivers stayed INERT with their blocking fields
named -- Whisper cross-attention rows, wav2vec2 CTC probability words, PCM
extrema. Authoring those numbers would have been a contaminated fixture,
which is worse than an inert driver.

The Python extractors ARE the intended upstream, answered by checking rather
than assuming: espeak-ng, ffmpeg, numpy and network are present and BOTH
checkpoints are already cached; torch, transformers and soundfile are not.
One pip install plus a measurement run away -- and taking a measurement is a
different act from recovering a file.

## A cost law, measured

MorphologicalLanguageEcology::generate branches super-exponentially in
maximum_generated_tokens. Release build, 23-passage corpus, one prompt:

    1 token       57 ms
    2 tokens     403 ms
    4 tokens  55,098 ms
    8 tokens  did not return in 200,000 ms

The laboratory's declared 256-token aperture is unreachable. Both the
faithful and the bounded sources ship, and the measurement is recorded in
the files' own generation_aperture_note so the bound can never be mistaken
for the law.

## The manifest was becoming the clutter it prevents -- my design flaw

One morphological run deposited 13,154 .form artifacts and took the tracked
manifest to 14,046 lines and 2.7 MB. A manifest that grows by thirteen
thousand lines per run is the pollution it exists to catch.

Rolled up: ONE LINE PER DRIVER, 35 lines total. sha256 is taken over the
sorted addresses of that driver's artifacts, so it moves if any byte of any
artifact moves; returns and bytes move if any artifact is added or removed.
No threshold and no aperture -- every driver directory is rolled up
uniformly. --files reconstructs the full listing on demand.

The roll-up loses nothing that matters, because the loss this file exists to
catch is a DIRECTORY VANISHING, not one artifact inside one.

27 drivers, 14,042 returns, 2,791,088,637 bytes, verifier clean.

## One canon collision, avoided rather than papered over

Naming the recovered directory observations/ collided with
THE_DOCUMENT_LAW section 1.8, which declares that genre absent in the live
tree, and resolve_named_paths enforces it. Renamed to driver-sources/ and
RESULTS.md to LABORATORY_RESULTS.md, content untouched; the contradiction
count returned to its pre-existing four, verified by moving the directory
aside.

Three recovered sources have no driver here and are kept labelled ownerless,
citing nothing.

1588 passed / 0 failed / 14 ignored. resolve_named_paths 0 failures.

+++ crates/holonic-engine/src/rebase_invariants.rs
@@ -298,7 +298,7 @@ pub struct ReductionWork {
 
 impl ReductionWork {
     fn record(&mut self, value: &BigInt) {
-        let bits = u64::from(value.bits());
+        let bits = value.bits();
         self.entries_written = self.entries_written.saturating_add(1);
         self.written_bits = self.written_bits.saturating_add(bits);
         if bits > self.peak_entry_bits {
